Move checkpoint serialisation to internal package
