+++ core-deterministic/build.gradle
@@ -50,6 +50,7 @@ task patchCore(type: Zip, dependsOn: coreJarTask) {
     from(zipTree(originalJar)) {
         exclude 'net/corda/core/internal/*ToggleField*.class'
         exclude 'net/corda/core/serialization/*SerializationFactory*.class'
+        exclude 'net/corda/core/serialization/internal/CheckpointSerializationFactory*.class'
     }
 
     reproducibleFileOrder = true
