+++ core/src/main/kotlin/net/corda/core/serialization/SerializationAPI.kt
@@ -233,19 +233,10 @@ enum class ContextPropertyKeys {
 @KeepForDJVM
 object SerializationDefaults {
     val SERIALIZATION_FACTORY get() = effectiveSerializationEnv.serializationFactory
-    val CHECKPOINT_SERIALIZATION_FACTORY get() = effectiveSerializationEnv.checkpointSerializationFactory
     val P2P_CONTEXT get() = effectiveSerializationEnv.p2pContext
     @DeleteForDJVM val RPC_SERVER_CONTEXT get() = effectiveSerializationEnv.rpcServerContext
     @DeleteForDJVM val RPC_CLIENT_CONTEXT get() = effectiveSerializationEnv.rpcClientContext
     @DeleteForDJVM val STORAGE_CONTEXT get() = effectiveSerializationEnv.storageContext
-    @Deprecated("Checkpoint context is no longer available from SerializationDefaults, but must be obtained from CheckpointSerializationDefaults instead.")
-    @DeleteForDJVM val CHECKPOINT_CONTEXT: SerializationContext get() =
-        throw UnsupportedOperationException("Checkpoint context is no longer available from SerializationDefaults, " +
-                "but must be obtained from CheckpointSerializationDefaults instead.")
-}
-
-object CheckpointSerializationDefaults {
-    @DeleteForDJVM val CHECKPOINT_CONTEXT get() = effectiveSerializationEnv.checkpointContext
 }
 
 /**
