--- node/src/main/kotlin/net/corda/node/serialization/kryo/KryoServerSerializationScheme.kt
@@ -1,4 +1,0 @@
-package net.corda.node.serialization.kryo
-
-//TODO: get rid of this entirely (but not yet, as it lives in the public API)
-class KryoServerSerializationScheme : AbstractKryoSerializationScheme()
