+++ node/src/main/kotlin/net/corda/node/services/statemachine/FlowStateMachineImpl.kt
@@ -12,8 +12,8 @@ import net.corda.core.cordapp.Cordapp
 import net.corda.core.flows.*
 import net.corda.core.identity.Party
 import net.corda.core.internal.*
-import net.corda.core.serialization.CheckpointSerializationContext
-import net.corda.core.serialization.checkpointSerialize
+import net.corda.core.serialization.internal.CheckpointSerializationContext
+import net.corda.core.serialization.internal.checkpointSerialize
 import net.corda.core.utilities.Try
 import net.corda.core.utilities.debug
 import net.corda.core.utilities.trace
